+++ CHANGELOG.md
@@ -120,3 +120,4 @@
 -   Inventory command.
 -   Research command.
 -   Shop BR command.
+-   Catalog manager.